chore: refactored package name again
